Make "internal" classes non-public
We don't use OSGi anymore, so there is no notion of internal packages.
Instead, we can put all the classes in the same packages and make the
ones we want to keep hidden package-visible.

Since the test classes are now in a package with the same name, they
have access to the non-public runtime classes directly! Ain't that
nifty.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>